+++ src/linkaddressbook/addressbook.h
@@ -79,7 +79,7 @@ namespace Konversation
              * Send the file to the contact
              */
             void sendFile(const QString &uid, const KUrl &sourceURL,
-                const QString &altFileName = QString(), uint fileSize = 0);
+                const QString &altFileName = QString(), quint64 fileSize = 0);
 
         public:
             /**
